api/requirements.txt
transformers==4.35.2
llama-cpp-python==0.2.6
Flask
torch